Refactoring endpoints (#59)
* Refactored endpoints to fp.py and pm.py
